requirements.txt
pandas
requests
beautifulsoup4
lxml
jupyter
python-dotenv
pytest